requirements.txt
python-vlc==3.0.20123
schedule==1.2.2
